+++ src/index.js
@@ -1,4 +1,12 @@
-const { debug, info, startGroup, endGroup, getInput, setFailed } = require('@actions/core');
+const {
+  debug,
+  info,
+  startGroup,
+  endGroup,
+  getInput,
+  setFailed,
+  warning
+} = require('@actions/core');
 const { join } = require('path');
 const { getAll, getList } = require('./repo');
 const { writeFileSync, existsSync, mkdirSync } = require('fs');
@@ -10,23 +18,40 @@ async function run() {
   debug(`repos_path: ${repos_path}`);
   var list_path = join(repo_list_cache, 'repo-name.json');
   debug(`list_path: ${list_path}`);
-  if (!existsSync(repo_list_cache)) mkdirSync(repo_list_cache);
   try {
     startGroup('Get input value');
     const user = getInput('user', { require: false });
     info(`[Info]: user: ${user}`);
     const maxPage = getInput('maxPage', { require: false });
     info(`[Info]: maxPage: ${maxPage}`);
+    const isDebug = getInput('debug', { require: false });
+    info(`[Info]: isDebug: ${isDebug}`);
+    if (!existsSync(repo_list_cache)) {
+      if (isDebug) mkdirSync(repo_list_cache);
+    } else {
+      if (isDebug) throw Error(`The cache directory(${repo_list_cache}) is occupied!`);
+      else
+        warning(
+          `[Warning]: The cache directory(${repo_list_cache}) is occupied! ` +
+            '\n' +
+            '[Warning]: If debug option set to be true, it will be Error!'
+        );
+    }
+    warning(
+      `[Warning]: The cache directory(${repo_list_cache}) is occupied! ` +
+        '\n' +
+        '[Warning]: If debug option set to be true, it will be Error!'
+    );
     endGroup();
 
     startGroup('Get repo list');
-    var repo_list = await getAll(user, maxPage);
-    writeFileSync(repos_path, JSON.stringify(repo_list, null, 2), 'utf-8');
+    var repo_list = await getAll(user, maxPage, isDebug);
+    if (isDebug) writeFileSync(repos_path, JSON.stringify(repo_list, null, 2), 'utf-8');
     var repo_name = await getList(repo_list);
-    writeFileSync(list_path, JSON.stringify(repo_name, null, 2), 'utf-8');
+    if (isDebug) writeFileSync(list_path, JSON.stringify(repo_name, null, 2), 'utf-8');
     endGroup();
   } catch (error) {
-    debug(`error[run]: ${error}`);
+    debug(`Error[run]: ${error}`);
     setFailed(error);
   }
 }